+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_coop_per_player_loadout/gwo_loadouts.js
@@ -8,13 +8,6 @@ function gwoLoadouts() {
   gwoLoadoutsLoaded = true;
 
   try {
-    // "Temporarily" until the dealing of its minions is fixed
-    const removeGeneralCommander = function (cards) {
-      return _.reject(cards, function (card) {
-        return card.id() === "gwc_start_subcdr";
-      });
-    };
-
     requireGW(
       [
         "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/loadouts.js",
@@ -22,9 +15,7 @@ function gwoLoadouts() {
         "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/deal.js",
       ],
       function (loadouts, GWInventory, gwoDeal) {
-        const startCards = removeGeneralCommander(loadouts.startCards);
-
-        model.startCards(startCards);
+        model.startCards(loadouts.startCards);
         model.gwoCards = gwoDeal.setupGwoCards();
 
         const cards = [];